feat: useBaseballGame custom hook 분리
